+++ NEWS.md
@@ -1,3 +1,10 @@
+## R package processR version 0.2.0.3
+=====================================
+(2019-May-18)
+
+- New function drawCatModel2 added
+
+
 ## R package processR version 0.2.0.2
 =====================================
 (2019-May-15)